Rearranging utility macros
